Resolve PR comments

+++ onnxruntime/core/providers/cpu/nn/shrink.h
@@ -11,11 +11,11 @@ class Shrink final : public OpKernel {
   Shrink(const OpKernelInfo& op_kernel_info) : OpKernel(op_kernel_info) {
     float bias_temp;
     ORT_ENFORCE(op_kernel_info.GetAttr<float>("bias", &bias_temp).IsOK());
-    bias_ = gsl::narrow_cast<float>(bias_temp);
+    bias_ = bias_temp;
 
     float lambd_temp;
     ORT_ENFORCE(op_kernel_info.GetAttr<float>("lambd", &lambd_temp).IsOK());
-    lambd_ = gsl::narrow_cast<float>(lambd_temp);
+    lambd_ = lambd_temp;
   }
 
   Status Compute(OpKernelContext* p_op_kernel_context) const override;
